Fixed style issues
