+++ 3.0/HSSim/Sets/Basic/Mage/Spells/Flamestrike.cs
@@ -15,5 +15,4 @@ class Flamestrike : Spell
             return new SingleSubBoardContainer(cln, b, "Play Flamestrike");
         });
     }
-       
 }
