+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/BoulderfistOgre.cs
@@ -1,9 +1,6 @@
-﻿using System;
-
-class BoulderfistOgre : Minion
+﻿class BoulderfistOgre : Minion
 {
     public BoulderfistOgre() : base(6, 6, 7)
     {
-
     }
 }
