+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/CoreHound.cs
