+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/DragonlingMechanic.cs
@@ -1,6 +1,4 @@
-﻿
-
-class DragonlingMechanic : BattlecryMinion
+﻿class DragonlingMechanic : BattlecryMinion
 {
     public DragonlingMechanic() : base(4, 2, 4)
     {
