+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/FrostwolfGrunt.cs
@@ -1,6 +1,4 @@
-﻿
-
-class FrostwolfGrunt : Minion
+﻿class FrostwolfGrunt : Minion
 {
     public FrostwolfGrunt() : base(2, 2, 2)
     {
