+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/MurlocRaider.cs
@@ -1,9 +1,6 @@
-﻿using System;
-
-class MurlocRaider : Minion
+﻿class MurlocRaider : Minion
 {
     public MurlocRaider() : base(1, 2, 1)
     {
-
     }
 }
