+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/Nightblade.cs
