+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/NoviceEngineer.cs
