+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/StonetuskBoar.cs
@@ -1,6 +1,4 @@
-﻿
-
-class StonetuskBoar : Minion
+﻿class StonetuskBoar : Minion
 {
     public StonetuskBoar() : base(1, 1, 1)
     {
