+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/StormpikeCommando.cs
@@ -1,6 +1,4 @@
-﻿
-
-class StormpikeCommando : BattlecryMinion
+﻿class StormpikeCommando : BattlecryMinion
 {
     public StormpikeCommando() : base(5, 4, 2)
     {
