+++ 3.0/HSSim/Sets/Basic/Neutral/Minions/StormwindChampion.cs
