+++ 3.0/HSSim/Sets/Basic/Neutral/Tokens/Boar.cs
@@ -1,6 +1,4 @@
-﻿
-
-class Boar : Minion
+﻿class Boar : Minion
 {
     public Boar() : base(1, 1, 1)
     {
