+++ 3.0/HSSim/Sets/Basic/Neutral/Tokens/MechanicalDragonling.cs
@@ -1,6 +1,4 @@
-﻿
-
-class MechanicalDragonling : Minion
+﻿class MechanicalDragonling : Minion
 {
     public MechanicalDragonling() : base(1, 2, 1)
     {
