+++ 3.0/HSSim/Sets/Basic/Neutral/Tokens/MurlocScout.cs
@@ -1,6 +1,4 @@
-﻿
-
-class MurlocScout : Minion
+﻿class MurlocScout : Minion
 {
     public MurlocScout() : base(1, 1, 1)
     {
