+++ 3.0/HSSim/Sets/Basic/Paladin/Tokens/SilverHandRecruit.cs
@@ -1,9 +1,6 @@
-﻿
-
-class SilverHandRecruit : Minion
+﻿class SilverHandRecruit : Minion
 {
     public SilverHandRecruit() : base(1, 1, 1)
     {
-
     }
 }
