+++ 3.0/HSSim/Sets/Basic/Rogue/Tokens/WickedKnife.cs
@@ -1,9 +1,6 @@
-﻿
-
-class WickedKnife : Weapon
+﻿class WickedKnife : Weapon
 {
     public WickedKnife() : base(1, 1, 2)
     {
-
     }
 }
