+++ 3.0/HSSim/Sets/Basic/Shaman/Tokens/WrathOfAirTotem.cs
@@ -1,6 +1,4 @@
-﻿
-
-class WrathOfAirTotem : Minion
+﻿class WrathOfAirTotem : Minion
 {
     public WrathOfAirTotem() : base(1, 0, 2)
     {
