+++ 3.0/HSSim/Sets/Basic/Warrior/Weapons/FieryWarAxe.cs
@@ -1,9 +1,6 @@
-﻿
-
-class FieryWarAxe : Weapon
+﻿class FieryWarAxe : Weapon
 {
     public FieryWarAxe() : base(3, 3, 2)
     {
-
     }
 }
